Extract store
